+++ src/main/java/com/github/projetoleaf/repositories/TipoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.github.projetoleaf.beans.Tipo;
 
 @Repository
-public interface TipoRepository extends JpaRepository<Tipo, Integer> {
+public interface TipoRepository extends JpaRepository<Tipo, Long> {
 
 }
